Only increment generations for explicitly specified models
Instead of incrementing the generation on *every* database update and
maintaing a blacklist of when to skip, generations are now incremented
only when models explicitly added to JIMMY_PAGE_WATCHLIST are
saved/updated or deleted.

Also, specify models with the Django-esque `app.Model` syntax instead
of the more tedious tablename.